Re-write translation framework to work with WebKit 2.
Opted for a simpler approach - use HTML directly from the `i18n` folder.

This validates if translations exist for locale and the HTML files are
present, then loads that page, which is the same as the original but
with generated translations.

Other changes ensures Welcome corrects paths to ../../ if being served
from a i18n folder.