Add a "pretty" repository listing feature.
Can be found in Boutique's preferences under the "Advanced" line.

Intended to allow a curious user to understand how their system sources
repositories. It would also be useful for diagnosis when users ask for
help in the community.

Also adds a "Known Repositories" area to the index. PPAs / sources
should be listed here that the Boutique are aware about.