Use GTK theme colours.
The theming is static once the program has initialized.

This is updated when navigating between pages, and possibly after
changing the theme on the Colour Selection page.

Some changes (like buttons) are not persistent with the rest of the OS
since Bootstrap styling is still used. Tested and working great against
the current Ubuntu MATE/MATE themes.